Send field contents as a name-val array to zemcontact.deliver
Opens up the possiblity of integrating ZCR with mailing list solutions such as MailChimp. A simple plugin that hooks into the callback could read the user-submitted content, a hidden 'list id' field and fire it all off along with the API key to subscribe someone to the list, then return a 'skip' notification so ZCR completes with success, but without doing any actual mailing.